requirements.txt
openai
streamlit
pandas
pytz
python-dotenv
pyairtable